+++ modules/container-web-app/datasources.tf
@@ -10,16 +10,6 @@ data "azurerm_application_insights" "app_insights" {
   resource_group_name = split("/", var.application_insights_id)[4]
 }
 
-data "azurerm_app_service_certificate" "certificate" {
-  for_each = var.custom_domains != null ? {
-    for k, v in var.custom_domains :
-    k => v if try(v.certificate_id != null, false)
-  } : {}
-
-  name                = reverse(split("/", each.value.certificate_id))[0]
-  resource_group_name = var.resource_group_name
-}
-
 module "backup_sas_token" {
   source  = "claranet/storage-sas-token/azurerm"
   version = "2.0.0"